backend/requirements.txt
aiosmtplib==2.0.0
alembic==1.8.1
anyio==3.6.1
apturl==0.5.2
arabic-reshaper==2.1.3
asgiref==3.6.0
async-generator==1.10
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.2
black==22.12.0
blinker==1.5
Brlapi==0.8.3
Brotli==1.0.9
certifi==2022.9.14
cffi==1.15.1
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==37.0.4
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.1.5
django-allauth==0.50.0
django-ckeditor==6.5.0
django-cleanup==6.0.0
django-js-asset==2.0.0
dnspython==2.2.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
email-validator==1.3.0
exceptiongroup==1.1.0
fastapi==0.85.0
fastapi-mail==1.2.4
fasteners==0.14.1
filelock==3.9.0
future==0.18.2
GDAL==3.4.3
gpg===1.16.0-unknown
greenlet==1.1.3
gunicorn==20.1.0
gyp==0.1
h11==0.14.0
httpcore==0.16.3
httplib2==0.20.2
httpx==0.23.3
identify==2.5.15
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==0.4.3
natsort==8.0.2
netifaces==0.11.0
nodeenv==1.7.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
outcome==1.2.0
packaging==21.3
paramiko==2.9.3
passlib==1.7.4
pathspec==0.10.3
pexpect==4.8.0
Pillow==9.2.0
platformdirs==2.6.2
pre-commit==3.0.0
protobuf==3.12.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.2
PyGObject==3.42.1
PyJWT==2.4.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0
reportlab==3.6.8
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
SecretStorage==3.3.1
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==1.4.41
sqlparse==0.4.3
starlette==0.20.4
systemd-python==234
text-unidecode==1.3
texttable==1.6.4
tomli==2.0.1
torbrowser-launcher==0.3.3
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.3.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.12
usb-creator==0.3.7
uvicorn==0.20.0
virtualenv==20.17.1
wadllib==1.3.6
websocket-client==1.2.3
wsproto==1.2.0
wxPython==4.0.7
xdg==5
xkit==0.0.0
zipp==1.0.0
